+++ serviceTest/src/test/scala/org/broadinstitute/dsde/workbench/dao/Google.scala
@@ -18,10 +18,12 @@ object Google {
   val ec: ExecutionContextExecutor = ExecutionContext.global
 
   val pemMode = GoogleCredentialModes.Pem(WorkbenchEmail(ServiceTestConfig.GCS.qaEmail),
-                                          new File(ServiceTestConfig.GCS.pathToQAPem))
+                                          new File(ServiceTestConfig.GCS.pathToQAPem)
+  )
   val pemModeWithServiceAccountUser = GoogleCredentialModes.Pem(WorkbenchEmail(ServiceTestConfig.GCS.qaEmail),
                                                                 new File(ServiceTestConfig.GCS.pathToQAPem),
-                                                                Option(WorkbenchEmail(ServiceTestConfig.GCS.subEmail)))
+                                                                Option(WorkbenchEmail(ServiceTestConfig.GCS.subEmail))
+  )
 
   lazy val googleIamDAO = new HttpGoogleIamDAO(appName, pemMode, metricBaseName)(system, ec)
   def googleBigQueryDAO(authToken: AuthToken): HttpGoogleBigQueryDAO =